factor out shell commands and final table helper
